% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffs.R
\name{categorical_cutoff}
\alias{categorical_cutoff}
\title{Categorical cutoff
Determines the opital cutoff for groups with numerical data according to
a chisquared statistics}
\usage{
categorical_cutoff(markers, values, strata, sig = 0.05, minN = 5, ...)
}
\arguments{
\item{values}{the values of the markers}

\item{strata}{the grouping}

\item{sig}{significance level (Default = 0.05)}

\item{minN}{minimun number per group (Default = 5)}

\item{makers}{the different markers to be analyses}
}
\description{
Categorical cutoff
Determines the opital cutoff for groups with numerical data according to
a chisquared statistics
}
\examples{
markers <- sample(letters[1:3], 100, replace = T)
values <- c(rnorm(50), rnorm(50, mean = 3))
strata <- c(rep("A", 50), rep("B", 50))
categorical_cutoff(markers, values, strata)
}
