% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{cluster_heatmap}
\alias{cluster_heatmap}
\title{Create}
\usage{
cluster_heatmap(d, n, show = T)
}
\arguments{
\item{d}{Tibble with an ID column, a key column and a value column}

\item{n}{number of clusters}
}
\description{
Create
}
\examples{

}
