% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compare_pearson_r}
\alias{compare_pearson_r}
\title{Compare two pearson correlation coefficients}
\usage{
compare_pearson_r(r1, r2, n1, n2)
}
\arguments{
\item{r1}{First pearson correlation}

\item{n2}{}
}
\description{
Compare two pearson correlation coefficients
}
