% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{create_tibble}
\alias{create_tibble}
\title{Create a tibble with random values}
\usage{
create_tibble(cols = 5, rows = 100, add_id = TRUE, type = "runif",
  verbose = F, debug = F)
}
\arguments{
\item{debug}{}
}
\description{
Create a tibble with random values
}
\examples{
# Normal distributed values
create_tibble(rows = 100, type = "rnorm")
}
