% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{DATA}
\alias{DATA}
\title{DATA
Wrapper for data path}
\usage{
DATA(name, verbose = T)
}
\arguments{
\item{name}{name of the file}

\item{verbose}{Verbose printing}
}
\description{
DATA
Wrapper for data path
}
