% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr_summary}
\alias{dplyr_summary}
\title{Create a summary like dataframe with piping}
\usage{
dplyr_summary(data, col)
}
\arguments{
\item{data}{data frame to be summarised}

\item{col_name}{the columns to be included}
}
\description{
Create a summary like dataframe with piping
}
\examples{
dplyr_summary(mtcars,cyl)
}
