% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorHandler.R
\name{errorWarrior}
\alias{errorWarrior}
\title{Error Hanlder}
\usage{
errorWarrior(x, abort_on_error = F, interactive = F)
}
\arguments{
\item{x}{Expression to be evaluated}

\item{abort_on_error}{should the execution be stoped upon an error (raises a proper Error)}

\item{interactive}{should you be prompted to enter an interactive browser session}
}
\description{
Error Hanlder
}
\examples{
errorHandler(prin("a"))
errorHandler(prin(a))
}
