% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_colors.R
\name{google_colors_accents}
\alias{google_colors_accents}
\title{Google color accents}
\usage{
google_colors_accents(level = 4, shuffle = F, show = F,
  distinct = F, n = NA)
}
\arguments{
\item{level}{Choose the level of saturation 1-4}

\item{shuffle}{shuffle the colors takes T or F}

\item{show}{show plot of the colors}

\item{distinct}{select distinct colors}
}
\value{
A vector of colors drawn from the google color pallets
}
\description{
Google color accents
}
\examples{
g <- google_colors_accents(level = 1, show = TRUE)
print(g)
}
