% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_colors.R
\name{google_colors_load}
\alias{google_colors_load}
\title{Load Google colors}
\usage{

  google_colors_load(url = "https://raw.githubusercontent.com/joelgsponer/waRRior/master/google/colors/waRRior.google.colors.json")
}
\arguments{
\item{url}{}
}
\description{
Load Google colors
}
