% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_extract_cluster}
\alias{heatmap_extract_cluster}
\title{Extract the clusters from a ComplexHeatmap}
\usage{
heatmap_extract_cluster(heatmap_obj, matrix_obj, which = "column")
}
\arguments{
\item{heatmap_obj}{a Heatmap object}

\item{matrix_obj}{the matrix that was used to create the heatmap, needs col and or row names}

\item{which}{row or column cluster}
}
\description{
Extract the clusters from a ComplexHeatmap
}
\examples{
mat = matrix(rnorm(80, 2), 8, 10)
mat = rbind(mat, matrix(rnorm(40, -2), 4, 10))
mat = matrix(rnorm(80, 2), 8, 10)
mat = rbind(mat, matrix(rnorm(40, -2), 4, 10))
rownames(mat) = letters[1:12]
colnames(mat) = letters[1:10]
HM <- ComplexHeatmap::Heatmap(mat, km=3)  #Make a heatmap, and have 3 clusters
HM <- ComplexHeatmap::draw(HM)  #Show the heatmap
heatmap_extract_cluster(HM, mat, which = "row")
}
