% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{IMG}
\alias{IMG}
\title{IMG
Wrapper for img path}
\usage{
IMG(name, verbose = T)
}
\arguments{
\item{name}{Name of the file}

\item{verbose}{Verbose printing (logical)}
}
\description{
IMG
Wrapper for img path
}
