% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute}
\alias{impute}
\title{Impute matrix}
\usage{
impute(d_mat, col_max_NA = 4, verboseFlag = F, remove_collinear = T,
  diagnostics = T)
}
\arguments{
\item{d_mat}{matrix to be imputed}

\item{col_max_NA}{maximum number of NA values per column. Columns with more will be removed}

\item{verboseFlag}{Vebose printing?}

\item{remove_collinear}{T/F}

\item{diagnostics}{T/F}
}
\value{
an imputed matrix
}
\description{
Impute matrix
}
