% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaked_functions.R
\name{max_noNA_noInf}
\alias{max_noNA_noInf}
\title{No NA -Inf Max
Returns a max value ignoring NAs while calculating the max and replacing
-Inf with NA (which occurs if the max is taken from only NA vales while
setting na.rm = T)}
\usage{
max_noNA_noInf(x)
}
\arguments{
\item{x}{}
}
\value{
Either a Number of NA
}
\description{
No NA -Inf Max
Returns a max value ignoring NAs while calculating the max and replacing
-Inf with NA (which occurs if the max is taken from only NA vales while
setting na.rm = T)
}
