% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaked_functions.R
\name{median_noNA_noInf}
\alias{median_noNA_noInf}
\title{No NA -Inf median
Returns a median value ignoring NAs while calculating the median and replacing
-Inf with NA (which occurs if the median is taken from only NA vales while
setting na.rm = T)}
\usage{
median_noNA_noInf(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Either a Number of NA
}
\description{
No NA -Inf median
Returns a median value ignoring NAs while calculating the median and replacing
-Inf with NA (which occurs if the median is taken from only NA vales while
setting na.rm = T)
}
\examples{
median_noNA_noInf(c(NA,1))
median_noNA_noInf(c(NA,NA))
}
