% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{message_glue}
\alias{message_glue}
\title{Message glue
Wrapper arount message and glue. Prints in green}
\usage{
message_glue(mes)
}
\arguments{
\item{mes}{}
}
\description{
Message glue
Wrapper arount message and glue. Prints in green
}
