% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaked_functions.R
\name{min_noNA_noInf}
\alias{min_noNA_noInf}
\title{No NA -Inf Min
Returns a min value ignoring NAs while calculating the min and replacing
-Inf with NA (which occurs if the min is taken from only NA vales while
setting na.rm = T)}
\usage{
min_noNA_noInf(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Either a Number of NA
}
\description{
No NA -Inf Min
Returns a min value ignoring NAs while calculating the min and replacing
-Inf with NA (which occurs if the min is taken from only NA vales while
setting na.rm = T)
}
\examples{
min_noNA_noInf(c(1,2))
min_noNA_noInf(c(NA,1))
min_noNA_noInf(c(NA,NA))
}
