% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Survival plot ggplot style}
\usage{
plot_survival(f, ...)
}
\arguments{
\item{f}{a survfit or summary.survfit object}

\item{...}{other parameters}
}
\description{
Survival plot ggplot style
}
