% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Title}
\usage{
plot_tsne(m, p = 30, t = 0.5, d = 2, clusters = NULL,
  color_clusters = F, which_clusters = "column",
  check_duplicates = F, point_size = 0.25, path = NULL, ...)
}
\arguments{
\item{m}{matrix}

\item{p}{perplexity}

\item{t}{theta}

\item{d}{dimensions}

\item{clusters}{A tibble or data frame with ID and Cluster}

\item{color_clusters}{Should the clusters be colored}

\item{which_clusters}{Column of row (for waRRior_heatmap)}

\item{path}{Provide a path to save the plot}

\item{...}{}
}
\value{
Returns a ggplot
}
\description{
Title
}
\examples{
# Without clusters
a <- runif(100)
b <- runif(100)
c <- runif(100)
m <- as.matrix(cbind(a,b,c))
plot_tsne(m)

# With clusters

a <- runif(100)
b <- runif(100)
c <- runif(100)
m <- as.matrix(cbind(a,b,c))
rownames(m) <- seq(1,100)
clusters = data.frame(ID = rownames(m), Cluster = sample(seq(1:4), size = 1000, replace = T))
plot_tsne(m, color_clusters = T, clusters = clusters)
}
