% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{print_if_verbose}
\alias{print_if_verbose}
\title{Verbose printig}
\usage{
print_if_verbose(message, verbose = F)
}
\arguments{
\item{message}{The messsage that should be printed in the console}

\item{verbose}{Prints to console only if T}
}
\description{
Prints the message to the console if the verbose parameter has
  been set to TRUE.
}
\examples{
print_if_verbose("Hello this is a verbose message", TRUE)
}
