% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_add_src}
\alias{project_add_src}
\title{Add a new src file
This file will automatically be soruced and added to the setup file}
\usage{
project_add_src(name, verbose = F, debug = F)
}
\arguments{
\item{name}{name of the file and function (prefixed with 'src_')}

\item{verbose}{Verbose printing}

\item{debug}{Debug printing}
}
\description{
Add a new src file
This file will automatically be soruced and added to the setup file
}
