% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_create}
\alias{project_create}
\title{Set up a project}
\usage{
project_create(verbose = F, debug = F)
}
\arguments{
\item{verbose}{Verbose printing}

\item{debug}{Debug printing}
}
\description{
Set up a project
}
