% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_setup_files}
\alias{project_setup_files}
\title{Setup files}
\usage{
project_setup_files(verbose = F, debug = F)
}
\arguments{
\item{verbose}{Verbose printing}

\item{debug}{Debug printing}
}
\description{
Setup files
}
