% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_setup_structure}
\alias{project_setup_structure}
\title{Setup project structure}
\usage{
project_setup_structure(verbose = T, debug = F)
}
\arguments{
\item{verbose}{Verbose printing}

\item{debug}{Debug printing}
}
\description{
Setup project structure
}
