% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_use_library}
\alias{project_use_library}
\title{Add library to be used in project}
\usage{
project_use_library(lib, file = "./src/lib.R", verbose = F,
  debug = F)
}
\arguments{
\item{lib}{The library to be added}

\item{file}{Where to add}

\item{verbose}{For verbose printing}

\item{debug}{For debug printing}
}
\description{
Add library to be used in project
}
