% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{reverse_factor_levels}
\alias{reverse_factor_levels}
\title{Reverse factor levels}
\usage{
reverse_factor_levels(x)
}
\arguments{
\item{x}{}
}
\value{
vector as a factor
}
\description{
Reverse factor levels
}
