% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaked_functions.R
\name{sd_noNA_noInf}
\alias{sd_noNA_noInf}
\title{No NA -Inf sd
Returns a sd value ignoring NAs while calculating the sd and replacing
-Inf with NA (which occurs if the sd is taken from only NA vales while
setting na.rm = T)}
\usage{
sd_noNA_noInf(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Either a Number of NA
}
\description{
No NA -Inf sd
Returns a sd value ignoring NAs while calculating the sd and replacing
-Inf with NA (which occurs if the sd is taken from only NA vales while
setting na.rm = T)
}
\examples{
sd_noNA_noInf(c(NA,1))
sd_noNA_noInf(c(NA,NA))
}
