% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{self_update}
\alias{self_update}
\title{Self Update from Github
Self updates the package, acctually it reinstalls it.}
\usage{
self_update()
}
\description{
Self Update from Github
Self updates the package, acctually it reinstalls it.
}
\examples{
self_update()
}
