% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_project.R
\name{shiny_setup}
\alias{shiny_setup}
\title{Set up shiny app inside project}
\usage{
shiny_setup(verbose = T, debug = F)
}
\arguments{
\item{verbose}{Verbose printing}

\item{debug}{Verbose printing}
}
\description{
Set up shiny app inside project
}
