% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_project.R
\name{shiny_symlink}
\alias{shiny_symlink}
\title{Create symlinks for shiny}
\usage{
shiny_symlink(file = NULL, verbose = T)
}
\arguments{
\item{file}{file to linked, if null standard files will be linked}

\item{verbose}{Verbose printing}
}
\description{
Create symlinks for shiny
}
