% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippet_full_line_title}
\alias{snippet_full_line_title}
\title{Generates a 80 chr sequence to be used in documentation}
\usage{
snippet_full_line_title(input, filling, alignment, verbose = F,
  debug = F)
}
\arguments{
\item{input}{Central input, e.g. title}

\item{filling}{Chr to be used as filler}

\item{alignment}{Lower numbers more to the left}

\item{verbose}{For verbose printing}

\item{debug}{For Debuging}
}
\value{
A string of 80 chr
}
\description{
Generates a 80 chr sequence to be used in documentation
}
\examples{
snippet_full_line_title(input = "Hello", filling = "-", alignment = 40)
}
