% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippet_section}
\alias{snippet_section}
\title{Creates a section}
\usage{
snippet_section(title, description, ts, verbose = F, debug = F)
}
\arguments{
\item{title}{Section heading}

\item{description}{Section description}

\item{ts}{Add timestamp of creation (T/F)}

\item{verbose}{For verbose printing}

\item{debug}{For debug printing}
}
\value{
A section comment
}
\description{
To be uses as a snippet
}
\examples{
snippet_section(title = "1st Chapter",
 description = "Let the saga begin"
 , ts = TRUE, verbose = TRUE)
}
