% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippet_timestamp}
\alias{snippet_timestamp}
\title{Add a Timestamp for Documentation}
\usage{
snippet_timestamp(message)
}
\arguments{
\item{message}{What to be written}
}
\value{
A string with a timestamp for documentation
}
\description{
Add a Timestamp for Documentation
}
\examples{
snippet_timestamp("test")
}
