% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{SRC}
\alias{SRC}
\title{SRC
Wrapper for src path}
\usage{
SRC(name, verbose = T)
}
\arguments{
\item{name}{Name of the file}

\item{verbose}{Verbose printing (logical)}
}
\description{
SRC
Wrapper for src path
}
