% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{stacked_percentage_plot}
\alias{stacked_percentage_plot}
\title{Stacked Percentage Barplot}
\usage{
stacked_percentage_plot(d, predictive_variable, dependent_variable,
  path = NULL, show_plot = T, dataset = NULL, generate_path = F)
}
\arguments{
\item{generate_path}{}
}
\description{
Stacked Percentage Barplot
}
