% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaked_functions.R
\name{sum_units}
\alias{sum_units}
\title{Sum units
Takes a vector of units and sums them and returns a new unit vector with
the same units.}
\usage{
sum_units(x)
}
\arguments{
\item{x}{Unit vector}
}
\value{
Sum as a unit
}
\description{
Sum units
Takes a vector of units and sums them and returns a new unit vector with
the same units.
}
\examples{
library(ggplot2)
x <- unit(c(1,2,3,4), units = "cm")
sum_units(x)
}
