% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_warrior}
\alias{survival_warrior}
\title{Survival ROC}
\usage{
survival_warrior(time, event, predvar, plot_best = T, plot_mean = T,
  plot_median = T, plot_diagnostics = T, verbose = F, minN = 5,
  sig = 0.05, max_cutoffs = 20)
}
\arguments{
\item{time}{vector with the time}

\item{event}{vector with the events}

\item{predvar}{the vector with the predictive variables}

\item{plot_best}{should the best one be plotted}

\item{plot_mean}{should the median one be plotted}

\item{plot_diagnostics}{should a diagnostic plot be plotted}

\item{verbose}{vebose printin?}

\item{minN}{minimum n per group}

\item{sig}{significance level (default <= 0.05)}

\item{max_cutoffs}{maximum number of cutoffs}
}
\description{
Survival ROC
}
