% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{TBL}
\alias{TBL}
\title{TBL
Wrapper for tbl path}
\usage{
TBL(name, verbose = T)
}
\arguments{
\item{name}{Name of the file}

\item{verbose}{Verbose printing (logical)}
}
\description{
TBL
Wrapper for tbl path
}
