% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tma.R
\name{tidy_column_names}
\alias{tidy_column_names}
\title{Tidy column names}
\usage{
tidy_column_names(tbl, pattern = "[-() /:]", replacement = "_")
}
\arguments{
\item{replacement}{}
}
\description{
Tidy column names
}
