% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{cluster_heatmap}
\alias{cluster_heatmap}
\title{Create clustered Heatmap}
\usage{
cluster_heatmap(d, n, colorscheme = c(rev(google_colors$Blue$accent[4]),
  "black", google_colors$Red$accent[4]), square = F, show = T)
}
\arguments{
\item{d}{Tibble with an ID column, a key column and a value column}

\item{n}{number of clusters}

\item{colorscheme}{The Colorscheme for the heatmap}

\item{square}{Make the tiles square (TRUE or FALSE)}

\item{show}{Print plot}
}
\description{
Create clustered Heatmap
}
\examples{

}
