% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_colors.R
\name{google_colors_accents}
\alias{google_colors_accents}
\title{Google color accents}
\usage{
google_colors_accents(level = 4, shuffle = F, show = F)
}
\arguments{
\item{show}{}
}
\description{
Google color accents
}
