% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaked_functions.R
\name{mean_noNA_noInf}
\alias{mean_noNA_noInf}
\title{No NA -Inf Mean
Returns a mean value ignoring NAs while calculating the mean and replacing
-Inf with NA (which occurs if the mean is taken from only NA vales while
setting na.rm = T)}
\usage{
mean_noNA_noInf(x)
}
\arguments{
\item{x}{}
}
\value{
Either a Number of NA
}
\description{
No NA -Inf Mean
Returns a mean value ignoring NAs while calculating the mean and replacing
-Inf with NA (which occurs if the mean is taken from only NA vales while
setting na.rm = T)
}
