% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippet_section}
\alias{snippet_section}
\title{Creates a section}
\usage{
snippet_section(title, description, verbose = F, debug = F,
  filling = "-")
}
\arguments{
\item{title}{Section heading}

\item{description}{Section description}

\item{verbose}{For verbose printing}

\item{debug}{For debug printing}

\item{filling}{- or + or #}
}
\value{
A section comment
}
\description{
To be uses as a snippet
}
\examples{
snippet_section(
  title = "1st Chapter"
, description = "Let the saga begin"
, verbose = TRUE
, filling = "-"
)
}
