% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tma.R
\name{tidy_column_names}
\alias{tidy_column_names}
\title{Tidy column names}
\usage{
tidy_column_names(tbl, pattern = "[-() /:\%.?]", replacement = "_")
}
\arguments{
\item{tbl}{Table where collumn names should be tidied}

\item{pattern}{What charcters should be removed}

\item{replacement}{Repclament character}
}
\value{
The cleaned table
}
\description{
Tidy column names
}
\examples{
fruits <- c("Pear", "Apple", "n.a", "X")
veggies <- c("Cauliflower", "X", "Carrot", "n.a")
d <- data.frame(veggies,fruits)
d <- waRRior::tidy_missing_values(d, c("n.a", "X"), c("veggies", "fruits"))
}
